+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.0.2 - 2016-05-19
+
+* Added delete method to the client
+* Added getValue method to the client, that returns the casted value instead of an array 
+* Deprecating get method since version 2.0
+
 ## 1.0.1 - 2016-05-18
 
 * Fixed bug when sending integer data to DynamoDb not casted as string